requirements.txt
absl-py==1.0.0
aniso8601==7.0.0
asgiref==3.5.0
astunparse==1.6.3
autopep8==1.5.2
beautifulsoup4==4.9.3
branca==0.4.1
cached-property==1.5.2
cachetools==5.0.0
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.2.12
django-bootstrap==0.2.4
django-bootstrap4==22.1
django-crispy-forms==1.10.0
django-environ==0.4.5
django-extensions==3.1.5
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-material==1.10.0
flatbuffers==2.0
folium==0.11.0
gast==0.5.3
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.55.0
graphene==2.1.8
graphene-django==2.13.0
graphene-file-upload==1.2.2
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.44.0
gunicorn==20.0.4
h5py==3.6.0
httplib2==0.20.4
idna==2.9
importlib-metadata==2.1.3
Jinja2==2.11.2
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.3
MarkupSafe==1.1.1
material==0.1
mtcnn==0.1.1
mysqlclient==2.0.1
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
Pillow==7.1.2
promise==2.3
protobuf==3.19.4
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydot==1.4.2
PyDrive==1.3.1
pyparsing==3.0.7
python-decouple==3.3
pytz==2020.1
PyYAML==6.0
requests==2.23.0
requests-oauthlib==1.3.1
rsa==4.8
Rx==1.6.1
scikit-learn==1.0.2
scipy==1.7.3
silence-tensorflow==1.1.1
singledispatch==3.4.0.3
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
sqlparse==0.3.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
typing_extensions==4.1.1
Unidecode==1.1.1
uritemplate==4.1.1
urllib3==1.25.9
Werkzeug==2.0.3
whitenoise==5.1.0
wincertstore==0.2
wrapt==1.14.0
zipp==3.7.0